#410: Make the apigee-openlegacy tool deployable on Apigee X/hybrid

+++ tools/apigee-sackmesser/pipeline.sh
@@ -27,14 +27,14 @@ unset APIGEE_TOKEN
 
 # Test Sackmesser Deploy
 
-# Sackmesser X Deployment
-BASE_PATH="/sackmesser-example-x"
+echo "Sackmesser X Deployment"
+BASE_PATH_X="/sackmesser-example-x"
 
 sackmesser deploy \
   --googleapi \
   -d "$SCRIPT_FOLDER"/test/sackmesser-example \
   -n sackmesser-x-v0 \
-  -b "$BASE_PATH" \
+  -b "$BASE_PATH_X" \
   -t "$APIGEE_X_TOKEN" \
   -o "$APIGEE_X_ORG" \
   -e "$APIGEE_X_ENV"
@@ -42,17 +42,17 @@ sackmesser deploy \
 (cd "$SCRIPT_FOLDER"/test && \
   npm i --no-fund && \
   TEST_HOST="$APIGEE_X_HOSTNAME" \
-  TEST_BASE_PATH="$BASE_PATH" \
+  TEST_BASE_PATH="$BASE_PATH_X" \
   npm run test)
 
-# Sackmesser Edge Deployment
-BASE_PATH="/sackmesser-example-edge"
+echo "Sackmesser Edge Deployment"
+BASE_PATH_EDGE="/sackmesser-example-edge"
 
 sackmesser deploy \
   --apigeeapi \
   -d "$SCRIPT_FOLDER"/test/sackmesser-example \
   --name sackmesser-edge-v0 \
-  --base-path "$BASE_PATH" \
+  --base-path "$BASE_PATH_EDGE" \
   --username "$APIGEE_USER" \
   --password "$APIGEE_PASS" \
   --organization "$APIGEE_ORG" \
@@ -61,10 +61,10 @@ sackmesser deploy \
 (cd "$SCRIPT_FOLDER"/test && \
   npm i --no-fund && \
   TEST_HOST="$APIGEE_ORG-$APIGEE_ENV.apigee.net" \
-  TEST_BASE_PATH="$BASE_PATH" \
+  TEST_BASE_PATH="$BASE_PATH_EDGE" \
   npm run test)
 
-# Test Sackmesser Await
+echo  "Test Sackmesser Await"
 # (Test readiness of the deployed API proxy)
 sackmesser await --googleapi -t "$APIGEE_X_TOKEN" proxy "sackmesser-x-v0"
 
@@ -72,23 +72,31 @@ docker run apigee-sackmesser await --apigeeapi --username "$APIGEE_USER" \
   --password "$APIGEE_PASS" --organization "$APIGEE_ORG" \
   --environment "$APIGEE_ENV" proxy "sackmesser-edge-v0"
 
-# Test Sackmesser List
+echo "Test Sackmesser List"
 # (List all proxies to find the previously deployed API proxy)
 sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/apis" | grep "sackmesser-x-v0"
 
 docker run apigee-sackmesser list --apigeeapi -u "$APIGEE_USER" -p "$APIGEE_PASS" "organizations/$APIGEE_ORG/apis" | grep "sackmesser-edge-v0"
 
-# Test Sackmesser Export
+echo "Test Sackmesser Export (X)"
 sackmesser export --googleapi -o "$APIGEE_X_ORG" -t "$APIGEE_X_TOKEN"
-if [ ! -d  "$APIGEE_X_ORG/proxies/sackmesser-x-v0" ]; then
-  echo "export failed"
-  exit 1
-fi
-rm -rf "$APIGEE_X_ORG"
+grep "$BASE_PATH_X" "$SCRIPT_FOLDER/$APIGEE_X_ORG/proxies/sackmesser-x-v0/apiproxy/proxies/default.xml"
+grep "sackmesser-app" "$SCRIPT_FOLDER/$APIGEE_X_ORG/config/resources/edge/org/developerApps.json"
+grep "janedoe@example.com" "$SCRIPT_FOLDER/$APIGEE_X_ORG/config/resources/edge/org/developers.json"
+grep "SackMesserProduct1" "$SCRIPT_FOLDER/$APIGEE_X_ORG/config/resources/edge/org/apiProducts.json"
+grep "SackMesserProduct2" "$SCRIPT_FOLDER/$APIGEE_X_ORG/config/resources/edge/org/apiProducts.json"
 
-docker run --entrypoint /bin/bash apigee-sackmesser -c "sackmesser export --apigeeapi -u $APIGEE_USER -p $APIGEE_PASS -o $APIGEE_ORG && ls $APIGEE_ORG/proxies/sackmesser-edge-v0"
+rm -rf "${SCRIPT_FOLDER:?}/$APIGEE_X_ORG"
 
-# Test Sackmesser Clean
+echo "Test Sackmesser Export (Edge)"
+sackmesser export --apigeeapi -u "$APIGEE_USER" -p "$APIGEE_PASS" -o "$APIGEE_ORG"
+grep "$BASE_PATH_EDGE" "$SCRIPT_FOLDER/$APIGEE_ORG/proxies/sackmesser-edge-v0/apiproxy/proxies/default.xml"
+grep "sackmesser-app" "$SCRIPT_FOLDER/$APIGEE_ORG/config/resources/edge/org/developerApps.json"
+grep "janedoe@example.com" "$SCRIPT_FOLDER/$APIGEE_ORG/config/resources/edge/org/developers.json"
+grep "SackMesserProduct1" "$SCRIPT_FOLDER/$APIGEE_ORG/config/resources/edge/org/apiProducts.json"
+grep "SackMesserProduct2" "$SCRIPT_FOLDER/$APIGEE_ORG/config/resources/edge/org/apiProducts.json"
+
+echo "Test Sackmesser Clean (Edge)"
 sackmesser clean proxy sackmesser-edge-v0 --apigeeapi -u "$APIGEE_USER" -p "$APIGEE_PASS" -o "$APIGEE_ORG" --quiet
 proxiesremaining=$(sackmesser list --apigeeapi -u "$APIGEE_USER" -p "$APIGEE_PASS" "organizations/$APIGEE_ORG/apis")
 if [ "$(echo "$proxiesremaining" | jq 'map(select(. == "sackmesser-edge-v0"))')" != "[]" ];then
@@ -96,6 +104,7 @@ if [ "$(echo "$proxiesremaining" | jq 'map(select(. == "sackmesser-edge-v0"))')"
   exit 1
 fi
 
+echo "Test Sackmesser Clean (X)"
 docker run apigee-sackmesser clean proxy sackmesser-x-v0 --googleapi -t "$APIGEE_X_TOKEN" -o "$APIGEE_X_ORG" --quiet
 
 proxiesremaining=$(sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/apis")
@@ -103,3 +112,34 @@ if [ "$(echo "$proxiesremaining" | jq 'map(select(. == "sackmesser-x-v0"))')" !=
   echo "failed to delete proxy"
   exit 1
 fi
+
+echo "Delete Test App to test Re-Import Edge Export in X"
+testappId=$(sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/developers/janedoe@example.com/apps/sackmesser-app" | jq -r '.appId')
+
+sackmesser clean app "$testappId" --googleapi -t "$APIGEE_X_TOKEN" -o "$APIGEE_X_ORG" --quiet
+sackmesser clean developer "janedoe@example.com" --googleapi -t "$APIGEE_X_TOKEN" -o "$APIGEE_X_ORG" --quiet
+sackmesser clean product "SackMesserProduct1" --googleapi -t "$APIGEE_X_TOKEN" -o "$APIGEE_X_ORG" --quiet
+sackmesser clean product "SackMesserProduct2" --googleapi -t "$APIGEE_X_TOKEN" -o "$APIGEE_X_ORG" --quiet
+
+echo "Test Re-Import Edge Export in X"
+sackmesser deploy \
+  --googleapi \
+  -d "$SCRIPT_FOLDER/$APIGEE_ORG/config" \
+  -t "$APIGEE_X_TOKEN" \
+  -o "$APIGEE_X_ORG" \
+  -e "$APIGEE_X_ENV"
+
+sackmesser deploy \
+  --googleapi \
+  -d "$SCRIPT_FOLDER/$APIGEE_ORG/proxies/sackmesser-edge-v0" \
+  -t "$APIGEE_X_TOKEN" \
+  -o "$APIGEE_X_ORG" \
+  -e "$APIGEE_X_ENV"
+
+sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/apis" | grep "sackmesser-edge-v0"
+sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/developers" | grep "janedoe@example.com"
+sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/developers/janedoe@example.com/apps" | grep "sackmesser-app"
+sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/apiproducts" | grep "SackMesserProduct1"
+sackmesser list --googleapi -t "$APIGEE_X_TOKEN" "organizations/$APIGEE_X_ORG/apiproducts" | grep "SackMesserProduct2"
+
+rm -rf "${SCRIPT_FOLDER:?}/$APIGEE_ORG"
